refactor(scripts): move stylelint-related files into subfolder
As mentioned in the previous commit, let's keep the clutter out of the
top-level `lint()` implementation. (This is in keeping with how we
already keep all the JSP-related stuff in a separate directory too.)
